Add support for +,- in subscripts
There are two types of subscripts a_k where k is a variable.
Since we capture this type of subscripts, we should also capture
a_{k+1} etc. However, the actual identifier pattern in that case is
a_{?x}.
The other type of subscript is F_\mathrm{g}, which is unrelated to
F_\mathrm{c}. In that case the actual identifer names are F_\mathrm{g}
 and F_\mathrm{c} and not F_{?x}. The problem to autmatically distingish
 between those two cases is that people tend to forget to put the subscript
 to text mode to denote that the subscript is a abbreviation for a word and
 not a variable.